Documentation and methods for Money.java
Added some fixes to the logic and handling of accessor types in Money class. Switched cents to an integer and use (double) to cast to the value of when it's GT 100. Changed copy constructor to return a string to use the clone() method instead. Changed the gets to convert cents and the main c'tor that's called to use the get's for converted amounts.

Not yet tested from driver class